+++ varfish/static/js/flags_comments.js
@@ -29,7 +29,6 @@ function clickVariantBookmark() {
   var outerThis = $(this).closest(".bookmark").find(".variant-bookmark");
   var caseUuid = $(this).data("case");
   var cell = $(this).closest(".variant-row").find(".toggle-variant-details");
-  var row = dt.row($(this).closest(".variant-row"));
 
   // Get variant description from triggering bookmark icon.
   if (structural_or_small == "small") {
@@ -85,7 +84,9 @@ function clickVariantBookmark() {
         var variantRow = $(outerThis).closest(".variant-row");
         variantRow.removeClass("variant-row-positive variant-row-uncertain variant-row-negative variant-row-empty variant-row-wip");
         variantRow.addClass("variant-row-" + summarizeFlags(data));
-        loadVariantDetails(row, cell);
+        if (structural_or_small == "small") {
+          loadVariantDetails(dt.row($(this).closest(".variant-row")), cell);
+        }
       }).fail(function(xhr) {
         // failed, notify user
         alert("Updating variant flags failed");
@@ -108,7 +109,9 @@ function clickVariantBookmark() {
           var commentTag = $(outerThis).closest(".bookmark").find(".variant-comment");
           commentTag.removeClass("fa-comment fa-comment-o");
           commentTag.addClass("fa-comment");
-          loadVariantDetails(row, cell);
+          if (structural_or_small == "small") {
+            loadVariantDetails(dt.row($(this).closest(".variant-row")), cell);
+          }
         }).fail(function(xhr) {
           // failed, notify user
           alert("Adding comment failed");